campaign_analytics.py
import streamlit as st
import plotly.express as px
import plotly.graph_objs as go
import pandas as pd
from datetime import datetime, timedelta

def generate_campaign_report(campaign_results):
    """
    Generate a comprehensive report for an email campaign
    
    Args:
        campaign_results (dict): Campaign sending results
    
    Returns:
        pd.DataFrame: Detailed campaign analytics
    """
    df = pd.DataFrame(campaign_results['delivery_details'])
    
    # Campaign Overview
    st.subheader("📊 Campaign Performance Overview")
    col1, col2, col3 = st.columns(3)
    
    with col1:
        st.metric("Total Recipients", campaign_results['total_recipients'])
    
    with col2:
        st.metric("Success Rate", f"{campaign_results['success_rate']:.2f}%")
    
    with col3:
        st.metric("Campaign ID", campaign_results['campaign_id'])
    
    # Delivery Status Pie Chart
    st.subheader("📨 Delivery Status Distribution")
    status_counts = df['status'].value_counts()
    fig_pie = px.pie(
        values=status_counts.values, 
        names=status_counts.index,
        title="Email Delivery Status",
        hole=0.3
    )
    st.plotly_chart(fig_pie)
    
    # Detailed Delivery Logs
    st.subheader("📝 Detailed Delivery Logs")
    st.dataframe(df)
    
    return df

def campaign_tracking_dashboard():
    """
    Create a comprehensive campaign tracking dashboard
    """
    st.title("🚀 Email Campaign Analytics")
    
    # Simulated campaign data (Replace with actual database retrieval)
    campaigns = [
        {
            "id": "campaign-001",
            "date": datetime.now() - timedelta(days=7),
            "total_recipients": 500,
            "success_rate": 95.2
        },
        {
            "id": "campaign-002", 
            "date": datetime.now() - timedelta(days=14),
            "total_recipients": 750,
            "success_rate": 92.5
        }
    ]
    
    # Campaign Performance Line Chart
    df_campaigns = pd.DataFrame(campaigns)
    fig_campaign_trend = px.line(
        df_campaigns, 
        x='date', 
        y='success_rate', 
        title='Campaign Success Rate Over Time'
    )
    st.plotly_chart(fig_campaign_trend)
    
    # Campaign List
    st.subheader("📋 Recent Campaigns")
    campaign_df = pd.DataFrame(campaigns)
    st.dataframe(campaign_df)
    
    # Detailed Campaign Selection
    selected_campaign = st.selectbox(
        "Select a Campaign for Detailed Analytics", 
        [camp['id'] for camp in campaigns]
    )
    
    # Placeholder for detailed campaign analytics
    if selected_campaign:
        # In a real scenario, fetch campaign details from database
        st.subheader(f"Detailed Analytics for Campaign {selected_campaign}")
        # You would call generate_campaign_report() with actual campaign data

def main():
    campaign_tracking_dashboard()

if __name__ == "__main__":
    main()